scripts/requirements-ingest.txt
olefile>=0.47,<1.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
openpyxl>=3.1,<4.0
python-pptx>=1.0,<2.0
